VerifyChecksumDialog: s/QFileSystemModel/QDirModel/g
The former requires us to pick a root path before showing anything.
QDirModel is nicer in that respect, as it just works. This is important here, as
we use only one model instance in potentially multiple views which may be rooted
in completely different parts of the file system. QFileSystemModel would then
watch the whole filesystem. I expect QDirModel to be less greedy.

svn path=/branches/kdepim/enterprise4/kdepim/; revision=1104693