Change OpenPGP key defaults to match GnuPG's
This causes kleopatra to generate OpenPGP keys
with a Sign / Certify primary key and an Encryption
only subkey.

The advanced options still give you the option to
generate a SCE Primary key without subkeys.

This change has been requested by the GnuPG maintainer
and was discussed on the Gpg4win-devel mailing list:
http://lists.wald.intevation.org/pipermail/gpg4win-devel/2015-April/001474.html

REVIEW: 124180