Port from KUniqueapplication to QApplication
Includes the change to KAboutdata so that we
can directly use KAboutdata with QApplication.

KleopatraApplication::newInstance now takes the commandline
parser and an optional working directory so that it can
also be used from the KDBusService activation.

The option to function mapping now uses Qt instead of boost.

Options are now declared in kleopatra_options.h.

The --import option no longer depends on the systrayicon.

Using the deamon option no longer prevents a MainWindow
beeing shown from a later start. (This was implemented through
firstNewInstance and probably never made sense.)

Command line parse errors are now printed to stderr or
shown in a message box if the application is already running.