Disable parallelized tests in test_verify
The disabled tests randomly fail due to thread safety issues in libgpgme.
There is an upstream bug for this [1] that has a test case attached
which mimics the behavior of the testParallelVerifyAndKeyListJobs
test. Until this bug is fixed there is nothing Kleopatra can do
to solve the issue.

The comment in testParallelVerifyAndKeyListJobs:
"### Increasing 10 to 500 makes the verify jobs fail!"
also indicates that this is not a new problem but that it has
been avoided in the past.

1: https://bugs.gnupg.org/gnupg/issue2092