Do not prefix query with 0x and show prefix info.
This generalizes the lookupcertificatescommand again so that
it can be used with a generic query. Prefixing 0x this low was wrong
and is also not necessary at least with SKS Keyservers.

Accordingly don't show the information text about how adding
or leaving out the 0x prefix might change results. Only show
it in case a search for a keyid or fingerprint does not return
a result.