Add --search command line option.
This option runs the lookupcertificatescommand with the provided
search string. This basically opens the external keysearch dialog.

The option implements the KDE 3 behavior of kleopatras
--query --external but I found it easier to understand if It is
named like the GnuPG --search option and does not mix the
"open certificate details / search on keyserver" options.

Behavior of the --query option is not documented a bit more
detailed in the help string.

CCBUG: 175980