Add support for ECDSA and ECDHE key generation
This adds the same basic keygen support (without proper subkey
handling) which is currently possible through gpgme API and kleopatra
for the basic ECC support that is available in all 2.1.x versions.

In the future with EDDSA added we probably want a more generic
subkey management with less repetitive code.