Fix logic in file system add handling
The bug was that after starting with an empty gnupghome the key
cache was not refreshed properly. Because the keyrings did
not initially exist. The logic to add newly created files was
broken as it checked the directory name against whitelist and
blacklist before checking for new files.

The problem was exposed by the recent change to whitelisting files
instead of blacklisting files.