Deprivatize and rework layout for crypto op config
This moves the two tabs with two options each in a single
layout with groups. It also tries to find better wording
for the "Quick Mode" where I did not understand the meaning
and wording (what was a conflict when singing things supposed to
mean?)
With more descriptive options the need for what's this should
be reduced and the strings are thus removed to avoid uninformative
explanations.