Add new KeyModel based on KeyRearangeColumnsProxy
This adds a direct depdencency to KItemModels as we now use
KRearrangeColumnsProxyModel to hide the columns from the model
that we do not want to show. This is also the first step for a
configurable column selection and order for which the
RearangeColumnsProxy gives us the flexibility.
This also adds a new hidden summary column as the first
column hidden in the normal tabwidget.