Rework certificateselectionwidget to lineedit
Behaves now more like Address entry field, shares model between
line edits. Full certificateselection dialog available on
request.

After user feedback a single line edit with the certificatelineedit
for free entry certificate selection and an improved combobox
for preset certificates was decided to be better then a mushed
one for all widget. So we now use a combination of certificatecombobox
and certificatelineedit for certificate selection.

When adding / removing recipients this should behave like the
address entry in KMail.

Brings up certificatedetails when a key is found, otherwise shows
the old certificate selection dialog which allows you to add
mutliple certificates in one go so that there is no
functionality lost.