Remove splashscreen and replace it with overlay
The startup should usually be quick enough that there is no
need for a splashscreen. If it takes longer we now show an
overlay over the keylistview and disable the keylistview.

This leaves the splashscreen in the repo / installtion
for now as there might be other uses for it.

Maniphest Tasks: T2240