Ensure keycache is populated on key search
To be more robust with the new background keycache init we
now check if the keycache is initialized when a find
find function of the cache is used.

This removes the need for callers to check if it is initalized.
Fixes key resolving in Verify Results after rev. fe12ed2a broke
it.