Do not depend on GpgME directly
By using the newly introduced PubkeyAlgo (5.2.81) and
functions in GpgMEpp we can get rid of the direct dependency
to GpgME.

This removes the last direct use of gpgme.h in Kleopatra.
Now we only have the dependency to GpgME++ and so no more
problems when multiple GpgME++ Versions are avaialable
where one is installed into GpgME's prefix.