Add AutoDecryptVerifyFilesController and use it
The AutoDecryptVerifyFilesController creates tasks based on
classification and immediately starts to decrypt / verify stuff.

Together with the new KeyCache logic this creates results much more
quickly and with fewer clicks then the old wizard style verify.

It is only used if a GpgME version with working identify support
is used as the basis for GpgMEpp (GpgME 1.7.0). The old controller
can still be used and serves as a backup in case autodetection fails.
(Although it should not and would be a bug first.)

If the tasks have output the output location is set in the new result
dialog and the output files are only moved out of the temporary
directory when the user confirms that they should be saved. This
fixes a Usability problem that the old dialog asked the user
to overwrite files before it showed the signature state but might
break use cases where decryption of bulk data is done over the network
so this might be configurable in the future.

Maniphest Task: T3027