Workaround crash / assert due to progress overflow
This is not the fix I wanted to have as it still does
not show the real progress. But previously there also
wasn't any progress so this is not a regression and it
fixes a crash.

See BKO 365931 for an outline of a proper fix that actually
shows progress.

BUG: 319281
CCBUG: 365931