Fix startup on UI-Server errors
This fixes the start when GPA is running. The try / catch
block was way too large and included important initialisation
code that was not run when the Server.start threw an exception.

Also only show the error / selftest problem on Windows as
it's not important currently for users of other platforms.