Do not encode PGP Mail Addresses
It's specified in RFC 4880 (5.11) that the user id is UTF-8 encoded.
So OpenPGP implementations handle this and (according to Werner Koch)
the E-mail should not be IDNA encoded.
This is also similar to the behavior of GPA and the gnupg command
line. gpgsm and X509 in contrast do require it to be encoded.