Fix progress handling (for gnupg >= 2.1.15)
We finally have a solution for progress that works with Qt
data types and how kleopatra passes data to GnuPG. This
requires GpgME 1.7.0 and GnuPG 2.1.15 to work.

This also simplifies progress handling for that case as we
can finally use current and total values and GnuPG no longer
overflows in progress messages since 2.1.14.

BUG: 365931