Refactor and fix Smartcard handling code
In preparation for OpenPGP card support

- The check / eventcounter code does not work as this code assumed
so the polling for that was useless. (Kleopatra only detected
smartcards on launch before this commit)
- We now soely rely on the filesystem watcher to get notfied about
smartcard changes. This is also what GPA does.
- Card is now a new class to manage the information we have about
a smartcard. This will be more information for OpenPGP smartcards.
- We are now prepared for more Card Types (like OpenPGP) and not
only for Netkey cards.
- No more boost required.
- Fixed Serial number detection. This never worked because SERIALNO
is a status and not data.