Add update notification for Gpg4win
As Kleopatra is the main "App" for Gpg4win it has to take
responsibility of GUI for updates. This is state of the art
and was a long missing feature.

Update checks are done by GnuPG / Dirmngr to avoid the pitfalls
of TLS and update verification. This readds the gpg4win-compact
pic but only installs it under Windows.

This also bumps the version to 1.8.0 but Messagelib already
depends on that version and a previous commit expected the
gpgmepp_version header to work like it does in 1.8.0