Use Normal color for verify with no information
If we have no information it's too strong to warn. This is
especially discouraging users to sign + encrypt because an
encrypted only message looks better to others who don't have
a key with a high validity for you.

The difference between Neutral and Negative was also to small
in the current KColorScheme for us to work.