Assist user with certify after import
If a single pubkey is imported ask the user to certify it
otherwise the gnupg system will not know if its valid or not.
This also tries to explain the "certification" process a bit.

And it has a don't ask me again button for users who
don't want to be bothered ;-)