Improve decrypt verify result display
This improves the look and information of the result
status display when verifying files.

The keys are now also fetched through GPGME and not
over the keycache to ensure that tofu information is
correct (if this is used). This also fixes a Bug
because previously signings subkeys (like the one
used to sign this commit) were not handled by
kleopatra's sig key lookup.

Although it's a bugfix it should stay in master
because of the string changes.

BUG: 340144