Remove comment field from OpenPGP key generation
The comment part of OpenPGP user ids is nowadays seen as a misfeature.
There is no consensus about what kind of information to put there, and
worse, users may not be aware that this information is published as
part of the certificate.

Some more rationale for this can be found under:
https://debian-administration.org/users/dkg/weblog/97

BUG: 354513
Differential: D6416

Signed-off-by: Justus Winter <justus@gnupg.org>

Commit message slightly amended by
Andre Heinecke <aheinecke@intevation.de>