Change some more wording away from Certificate
Mostly for the newcertificatewizard. We use Key Pair or
key pair now if we can't avoid it.

Sometimes we use _public_ key to make it clear that it is
intended to be published.

The wording OpenPGP Certificate never gained traction and is
confusing to most users it's also not consistent in Kleo itself.

This also removes the placebo text entry field in the key creation
wizard as users were very confused by that and it does not really
do something.