Readd Version Suffix and bump version to 3.0
This will be released together with Gpg4win-3.0 and brings
the versions more in line to avoid confusion. Also we changed
a lot since 2.3.0. So this is also a good time to bump.

The Version suffix was recently removed but is important to show
the "bundle" Version of the Windows installer that includes
Kleopatra as this determines also the Version of all the dependencies.