Fix detached S/MIME signature verification
GPG cannot properly distinguish whether a file contains a detached
or an opaque S/MIME signature. So we now first try to check if a file
with the same base name as the S/MIME signature file exists and treat
the signature as detached. If no such file exists, we assume opaque
signature.

This change also re-orders the way the tasks are scheduled to make sure
that decryption happens before signature verification. For that we need
to "pipe" output from the decryption task into the verification task.
Since holding the decrypted content in-memory could be very costly in
case of a large file, this change introduces Crypto::Input that can
work on Crypto::FileOutput in a way that the FileOutput signals to the
Input that the file now exists.

Differential Revision: https://phabricator.kde.org/D8376