Fix potential race hiding keycacheoverlay
There might be an issue that the keycache is filled before
the signal is properly connected. To be robust this adds
a watchdog to check if the keycache is initialized.

This mainly a safeguard. The problem reported in the
Bug appears to be a GnuPG problem as kleopatra would
show keys in the background if the keycache was filled.

CCBUG: 381910