Add URL handling to smartcardwidget
This adds setting and viewing the pubkey URL.
Directly importing the URL like GPA does is not
supported because users may link to pages that
not directly contain a key, and a valid key link
with mimetype application/pgp-keys should be opened
by the Desktop system in Kleopatra. This would then
trigger the import.