Clarify profile config GUI element
Rename it to "Activate GnuPG Profile" to make it more
clear that it does not show the currently active profile.

For this reason also use an empty entry as the default entry.

We can't show which profile is currently active as multiple
profiles can be activated at once and combined. GnuPG also
provides no way to query this information.

GnuPG-Bug-Id: T3723