Fix raising on second call if session restored
This was broken with rev. 212ccad96 the old logic
was that if the call to new instance was the first and
kleo isSessionRestored then the new instance should
not raise the main window.

Without this once Kleopatra is started with the session
parameter it won't raise the main window on a second call.