Change working dir on Windows after startup
After handling the filename arguments on the command line
switch the working directory on Windows to prevent the
locking the directory from which Kleo was started
for deletion.

This is important for the file extension support
of Kleopatra on Windows.

This also unifies two return paths in newInstance to
a single one to avoid duplication.

BUG: 393447