Change icon to open selection dlg in lineedit
I've got feedback from several users that they were missing
the certificate selection dialog because they did not
see the "emblem-question" as clickable and so missed
the way how the dialog is opened.

The resource-group-add is a good fit IMO because the dialog
is mostly useful to add multiple recipients at once.