Handle no MDC for legacy algos gracefully
With GPGME 1.11.2 we can detect if an MDC error occured
because of an old algorithm. This is common enough that
we need a proper handling.

It is now detected, the user is shown a notification and
gets the option to "Force decryption" so that the
user could then ideally re-encrypt the file with a modern
algo.

GnuPG-Bug-Id: T4038