Fix lookup dlg by porting it to keytreeview
It was actually not noticed in bugs or issues but it was broken
in that it showed all available columns (but the weird ones were
hidden behind a horizontal scrollbar).
We now use the keytreeview with the rearrangecolumnsmodel.

This is important to use the long key id here because for keyserver
searches that one is really important.

GnuPG-Bug-ID: T3238