Another try to fix windows encoding mess
We now put GnuPG's output through a central conversion
function so that we have a place to modify when things
change.

The Problem is that on my western windows system GnuPG
gets CP 437 as GetConsoleOutputCP and prints in that codepage.
We get 0 as GetConsoleOutputCP and 1252 with GetACP.

The only thing that seemed to somehow match was GetOEMCP but
that might just be luck and it might still be broken in
other windows languages.

Using --display-charset utf8 does not work (even when fixing
the argument order) and is also not available for GPGSM.

BUG: 360364