Use windows pipe io device for gpgtar
This adds a Windows specific QIODevice for IPC
using Windows Create Process and pipes.
This class serves as an alternative to use QProcess on
Windows which event driven nature does not play well
in our threading and IPC model.

This class was written with gpgtar in mind and mostly
a reaction to multiple issues we had with QProcess
and gpgtar in GPGME on Windows. It was so hard to debug them
that we decided for a simple approach that gives us
full control.

As there are use cases streaming terrabytes through this
even the control of the buffer size is an advantage.

GnuPG-Bug-Id: T4332