Show embedded file name if it differs
If the embedded filename differs from the one kleopatra
uses as output filename we show the name now.

This is especially useful for recipients of PGP/partitioned
messages where attachments are just named "Attachment1.pgp"
and Kleopatra would not show that it was for example an
excel document originally.

We only show the name and not use it for security because
the filename is an attacker controlled string.

GnuPG-Bug-Id: T4390