Pretty up and remove nag in certify dialog
This removes the nagging "I have verified the fingerprint"
checkbox. If users do not verify they just check it anyway.
Better to add a sentence why the fingerprint is important.

This also removes the duplicated information of the
primary userid + keyid and removes the usage of the
technical OpenPGP term "UserID" from the heading.

More in line with modern Kleopatra design.