Redesign certify dialog and add remark support
Remarks are new in GPGME 1.14 they are signature annotations
that can be added to user id signatures.
For users this can be used to add search tags and other useful
information.
This enables some work flows for organizations to easily
group and sort their keys consistently.

Remarks are shown from any key that has full Ownertrust.

The new certify dialog now also works much more smoothly
and uses the "modern" secret key selection widget. So
that it is no longer a wizard and is simpler and much
faster to use.

Users no longer have to have click orgies when they
want to sign several keys with multiple user ids.

GnuPG-Bug-Id: T4734