Remove "default" indicator for adv. key sizes
It's a bit complicated what kleopatra uses as the default
key size as this was only configurable in kleopatra but
nowadays can be configured in GnuPG and we also have
ECC which could also be default. So rather then to
show a wrong default indicator we do not indicate default
anymore.

GnuPG-Bug-Id: T4988