Fix fingerprint and creation timestamp written to OpenPGP cards
The timestamp needs to be in UTC. QDateTime::fromSecsSinceEpoch()
uses local time by default, so we need to request UTC explicitly.
The wrong creation timestamp resulted in a different creation
time stored on the card than stored locally. And, since the creation
time is included in the calculation of the fingerprint, this also
resulted in different fingerprints stored on the card and locally.