Use a more modern approach to watch for smartcard removal
Use the new DEVINFO --watch command to watch for status changes of
smartcard devices instead of using a file system watcher to watch
certain files. As with the old approach we are only notified about
the removal of smartcards unless we/the user explicitly trigger/s a
rescan (with the SERIALNO --all command).

GnuPG-bug-id: 5066