Allow unicode GnuPG Homedir
This uses gpgconf to look up the homedir of GnuPG and we now
work internally with UTF-8 filenames which are passed through
gpgrt in the end to have Windows UTF-16 conversion. This
fixes a crash / assert when Kleopatra is run from a homedirectory
that contains a non 8 bit character on Windows.