Remove overviewpage from newkeywizard
This removes one more step when first starting Kleopatra, the
user is asked to review stuff like algorithm and key sizes etc.
Which a user should not need to know about. In the spirit of
not confusing users its better to just remove that. Very interested
users can "review" afterwards.

GnuPG-Bug-Id: T5181