Allow creation of CSRs for card keys of OpenPGP cards
The creation of CSRs for the signing key and the authentication key
is possible (i.e. supported by gpgsm). The creation of a CSR for the
encryption key fails (because signing the request with this key fails
with "Invalid ID").

The creation of CSRs may still fail for certain OpenPGP keys, e.g.
ed25519 keys do not seem to be supported by gpgsm ("Invalid digest
algorithm"). I'm not checking this in Kleopatra in case those keys will
later be supported in gpgsm.

GnuPG-bug-id: 5183