Permit empty passphrase for key by default
As this change might be controversial we both check if
"enforce-passphrase-constraints" is set in the GnuPG backend
and make this option configurable in kleopatrarc.

For most people a passphrase on a key is not really required
since the endpoint is either fully compromised, in which case
a keylogger can obtain the passphrase or the local storage
is secure. The wording "encrypt key with an unrecoverable
passphrase" should also further clarify what happens when
a passphrase is entered.

GnuPG-Bug-Id: T5181
GnuPG-Bug-Id: T3724
GnuPG-Bug-Id: T4346