Prevent process elevation on Windows
On any modern Windows (>=Vista) there is the concept of elevated
security privileges. That means that you can elevate a user
of the Administrator group to have Administrative privileges but
e.g. still the same home directory as the user. This can
lead to subtle issues when Kleopatra / GnuPG creates files when
running elevated and Kleo is then later run not elvated. So
we better check for that and exit early if thats the case.

GnuPG-Bug-Id: T5212