Fix display of tags in key details for secret keys
The problem is that the option secretOnly is incompatible with the
option includeSigs (aka key list mode Signatures) because gpgme does
"gpg --list-secret-keys" if secretOnly is true,
"gpg --check-sigs" if secretOnly is false, but signatures are requested,
"gpg --list-keys" if secretOnly is false and no signatures are requested.
See T3580.

This problem caused the tags to vanish from the key list for the users'
own keys as soon as they viewed the key details.

Using key list mode WithSecret instead of secretOnly fixes the problem.

Note: Adding key list mode Signatures is superfluous because the job
is already created with includeSigs.

GnuPG-bug-id: 5174