Fix win_get_user_name
If GetUserName is used and not GetUserNameEx the size paremter
is one wchar_t too large and includes another \O. So it's better
to use fromWcharArray without size so that it only uses the
terminated string. If the zero is included in the GnuPG
key params it aborts parsing at that point and creates
a key without subkeys.