Add 7bit replacement option in archivedefinition
gpgtar does not properly handle local 8 bit encoding
and interprets it internally as UTF-8 breaking the
folder name created by gpgtar. To avoid having wrong
encoding in the filenames we now pass it for the
unpack command with all special characters replaced
by underscore.

This is an ugly hack but better then to leave it broken
and this code is only used if configured in libkleopatrarc.

GnuPG-Bug-Id: T5226