Fix Windows UNC paths for file encryption
With Qt 5.13.2 (and still there in 5.15.0)
there was a regression that QTemporaryFile on Windows
UNC drives (unmounted Samba shares) have a bad filename even
though the file is correctly created. As the fix in Qt for
this is not trivial this adds a workaround to fix up the path name
and can probably be removed in a future version.

GnuPG-Bug-Id: T5216