Change error when running as admin to warning
This changes the recently introduced check for users running
Kleopatra with elevated permissions to a warning. There are
still Windows users that run Kleopatra without any privilege
seperation whatsover on their system, they run everything
with full Administrative privileges. So for them the change
blocked usage of Kleopatra. I was unaware that this is still
possible in modern Windows versions as this is very hidden
and requires the user to disable safety systems.

GnuPG-Bug-Id: T5248