Improve combined S/MIME and OpenPGP export
This removes the exportcertificatesdialog which required
the user to select filenames without any suggestion by opening
file dialogs for both OpenPGP and S/MIME.
Now we just query for two output files and take the PGP filename
as a name suggestion for the S/MIME filename. This is much
quicker and IMO less confusing then the other dialog which
started empty and required the user to understand what to
do next.

Additionally we save in the config which directory was choosen
for the last export so that we can reuse that directory the
next time a user exports.

GnuPG-Bug-Id: T5002