Add a first impl and widget for PKCS#15 cards
This adds an implementation based on an atos/cardos 5.3 card
with applications from R&S. PKCS#15 cards might still
behave differently.
For now this is read only, just showing some status of the
card to show in the GUI that GnuPG can work with the card
and if we have an OpenPGP certificate for it.

GnuPG-Bug-Id: T4876