Search PKCS#15 keys on ldap servers
This needs to be moved in the readerstatus thread later to properly
handle it. The GUI is not the best place for it. But for now
this works, we do a search when we cannot find the keygrip for
the inserted smartcard and if we find a single key we import it
through --recv-key