PKCS#15: Update readerstatus after import
The problem here that caused crashes was that we should not
store the Card as a member variable as the reader might delete
it. In that case we would then access freed memory regions
after the update. So now we only use the fingerprint to update
the sigkeywidget which is the important one and otherwise call
a ReaderStatus::update to update everything.