Create shadow key also for the smartcard's encryption key.
The shadow key is required to properly decrypt messages.  The test
case here is: A new user inserts her card, Kleo figures which are
signing and encryption key and shows that in the dialog.  The user
enters the address of the recipient, signs and encrypts, and sends off
the data.  Now when she wants to decrypt the message by herself, she
will get an error because GnuPG can't find the key due to the missing
shadow key for the encryption key.  This might work better with GnuPG
2.3 but for 2.2 this is very annoying.  GnuPG actually does the same
what this patch does but that does not kick in here.

The --no-data is just an optimization available with latest gpg-agent
version; older version would just ignore this.