Edit the keyserver option that is actually used by gpg/dirmngr
gpg prefers the deprecated keyserver option in gpg.conf over the
keyserver option in dirmngr.conf; therefore, we use/edit the deprecated
keyserver option if it is set (or if the new option doesn't exist,
i.e. for gpg < 2.1.9); otherwise, we use/edit the new option.

Additionally, disable the option in the UI if, for whatever reason,
there is no usable config entry.

GnuPG-bug-id: 5461