Print finalizing message on 100% progress
When the progress of a task collection reaches 100% but it
is not finished it means usually that it is still waiting for
data to write out for example an archive to be extracted.

This way we show some indication that we are actually still doing
something and a knight rider progress.