Append a compact release service version to Kleopatra's version number
This ensures that we have different version numbers for different
releases without switching completely to the release service version.
This fixes problems with building flatpaks because appstream-compose
errors out if there are multiple <release> tags with the same version
attribute in the AppStream data (org.kde.kleopatra.appdata.xml).

This approach was suggested by Albert Astals Cid with reference to
khelpcenter from where I borrowed some of the code.