W32: Change window behavior to always activate
This code enables a creative solution in Qt for Windows to allow
the application to always raise a window. This is important
because kleopatra is kept running and might not have foreground
window permissions when a user triggers an action through
the uiserver or kuniqueservice_win. So many windows were opened
in the background.

GnuPG-Bug-Id: T5533