Take the scroll area out of the tab order
Set the focus policy of the scroll area to NoFocus. I think it makes no
sense to be able to give keyboard focus to the scroll area. On the
contrary, it was really confusing because neither the screen reader nor
the display gave any hint about which part of the GUI now has focus.
Took me some time and some additional debug output to figure this out.

GnuPG-bug-id: 5535