Ensure correct setup of UI for default protocol
Decouple the update of the UI after a change of the protocol from
setting the protocol. This way we can initialize the protocol already
in-class and still setup the UI depending on the protocol without having
to rely on the uninitialized mCurrentProto being different from
GpgME::UnknownProtocol.

The initialization of mCurrentProto in-class also ensures that other
setup code uses the correct value.

GnuPG-bug-id: 5535