Move focus to next (or previous) recipient field when removing one
If the recipient field to remove has focus, then we move the focus to
the next recipient field or, if the field to remove is the last one, to
the previous recipient field. Since there are at least two (empty) fields,
if we remove the field, we don't have to consider the case that there are
no remaining fields.

GnuPG-bug-id: 5535