Do not crash if there are no keys in the keyring
If there are no keys in the keyring, then we check the visibility
of the requesters before the dialog is shown on screen, i.e. before
the dialog is visible, so that isVisible() returns false even for
those requesters that were not explicitly set to "not visible". What
we actually want to find is not the first visible widget, but the
first widget that is not explicitly hidden. Therefore, we need to check
for !isHidden() instead of isVisible().

GnuPG-bug-id: 5535