Disable sign option and encrypt options if needed keys are not available
If there are no secret keys in the keyring, then disable the "sign as"
option and the "encrypt to me" option. (The latter because we currently
require a secret key for the "encrypt to me" key.)
If there are no public keys in the keyring (and therefore also no secret
keys), then disable the "encrypt for others" option and check the
"encrypt with passphrase" option which is the only possible operation in
this case.

GnuPG-bug-id: 5535