Update the commit button only if the corresponding wizard page is shown
If one presses Return while a recipient input widget has focus, then the
wizard switches to the next page and the SignEncryptWidget emits an
operationChanged signal which triggers an update of the commit button.
This update needs to be ignored, so that the label text that is
cleared in initializePage() of the second page is not set again.

GnuPG-bug-id: 5535