Revert "Add widget parent"
This reverts commit a4a84b9a3b48d501af9dcdf607bc5fffb3385669.

I'm sorry, but this change breaks the tab order in the dialog. In fact,
it makes it impossible to reach any widgets except for the two dialog
buttons with the Tab key. I also don't see why passing a parent to the
constructor should be necessary. All widgets are added to a layout and
the layout takes care of setting the correct parent widget of the widgets.