Fix defaulting to last used directory when importing certificates
Use QFileDialog::getOpenFileNames() instead of
Kleo::FileDialog::getOpenFileNames(). The Kleo::FileDialog helpers
expects the ID of a directory and not the path of a directory.
Moreover, those helpers only cache the last used directory transiently
in memory which is of no use for us because we want to store this
information persistently.

GnuPG-bug-id: 5638