Ensure that finished() isn't called twice if both export jobs fail
This is achieved by removing the completed jobs from the list of jobs
just before checking if the list is empty. Previously, it could happen
that the second job is removed while the error message for the first job
is still being displayed and then finished() was called for both
completed jobs crashing Kleopatra.

GnuPG-bug-id: 5638