Use primary algo also for encryption subkey if not specified otherwise
Example: If the default-new-key-algo option is set to "brainpoolp256r1",
then use this curve for the primary key and the encryption subkey
(instead of using cv25519 for the subkey). Just specifying a curve name
may not be the correct syntax for the default-new-key-algo option, but,
in my opinion, the user's intention is pretty obvious.

GnuPG-bug-id: 5717