Do not change read-only settings when loading defaults
Use the generated setters to set the settings to their defaults
because the setters actually do nothing, if settings are read-only,
while KConfigSkeleton::setDefaults() resets the settings to their
defaults even if they are immutable.

With this change all settings on this page/widget are reset to their
defaults. Previously, some file operations settings were missing.

The change also makes sure that the comboboxes are correctly reset to the
defaults, even if the defaults someday aren't at index 0 anymore.

GnuPG-bug-id: 5791