Fix invocation of the load() slot from KCModule::showEvent()
For some reason the save() override was invoked instead of the load()
override. Apparently, Qt's dispatch gets confused if overridden virtual
slots are not marked as slots. Marking the overridden slots as slots
fixes this.

In particular, this bug caused the X.509 servers to be saved as soon as
the config dialog was opened.

Note to self: Virtual slots are evil. Use non-virtual slots calling
(protected) virtual member functions.

GnuPG-bug-id: 5801