Fix canceling of import command
If a job is canceled, then the result signal is still emitted. This
caused an assert (or a crash) because in doCancel() the list of jobs
was cleared, but in the slot connected to the result signal the job
was still looked up in the (empty) list of jobs.

To fix this we disconnect all connections and call the importResult() slot
manually for all canceled jobs.

GnuPG-bug-id: 5805