Request start of GnuPG agent if connecting to it fails
The DeviceInfoWatcher needs a running GnuPG agent, so that it can
watch for changes of the smartcard devices. Request the start of the
GnuPG agent if connecting to the agent failed for the first time
(since the start or since the last successful connection).

This fixes the problem that the GnuPG agent wasn't started if the
user's keyring was empty which made it impossible to use a fresh
GNUPGHOME with a smartcard.

Moreover, the DeviceInfoWatcher will trigger a restart of the GnuPG
agent if it was shut down by the user with the new "Restart Background
Processes" action.

GnuPG-bug-id: 5775