Improve initial focus in Advanced Settings dialog
By default, the tab has input focus initially. That doesn't appear very
useful, in particular if there is only a single tab as for OpenPGP. So,
instead we focus the first checked (and enabled) key type radio button
or the first enabled check box or, as last resort, the OK button.

GnuPG-bug-id: 5832