Disable input focus for UI element showing resulting user id/DN
The UI element showing the resulting user id/DN built from the values
of the other input fields is read-only. Allowing it to get input focus
allows the user to copy its text, but it doesn't allow editing the text
although is looks like it would. I think that this is more confusing than
helpful.

GnuPG-bug-id: 5832