Allow column-by-column navigation in certificate list
This overrides the default handling of left/right cursor movement of
QTreeView. The left/right keys always move the cursor to the previous/next
column (or move the horizontal scroll bar). QTreeView's expanding and
collapsing of items with children is not done. The user has to use Plus,
Minus, or Asterisk for expanding or collapsing items.

Navigation column by column is important for people using screen
readers because this way they can get the values of the individual
entries of the table. Visually it may not be possible to see which entry
is the current entry because we always select complete rows, but that
shouldn't be a problem because it doesn't matter which entry of a row is
the current entry since there is no interaction possible with individual
table entries. Any interaction always applies to the selected rows.

GnuPG-bug-id: 5841