Simplify the action of the recipient key input field
Make the action of the recipient key input field always trigger the
details dialog of the current certificate or group (if there is one).
Note: This action is not accessible via keyboard or for screen readers,
but that's okay because the details dialog can also be opened via the
context menu.

Moreover, use only emblem-* icons for the action to give the different
statuses a unified look.

GnuPG-bug-id: 5845